Add some functionality tests for hfst-proc
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@624 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/hfst-check-help.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 for f in ../../tools/src/hfst-* ; do
-    if [ -x "$f" ] ; then
+    if [ -x "$f" -a ! -d "$f" ] ; then
         if ! "$f" --help > help.out ; then
             rm help.out
             echo $f has broken help
